@@ -1,9 +1,8 @@
 package com.example.yukai.mydefinedview.ArtLearningNote.chapter12_1;
 
 import android.app.Activity;
-import android.graphics.BitmapShader;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
