@@ -2,7 +2,7 @@ package com.example.yukai.mydefinedview.ArtLearningNote.chapter13_1;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.Button;
 
