@@ -2,7 +2,7 @@ package com.example.yukai.mydefinedview.ArtLearningNote.chapter3_2;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.Log;
 import android.view.MotionEvent;
 import android.view.View;
