@@ -2,8 +2,8 @@ package com.example.yukai.mydefinedview.BasicView.FragmentAndActivityTest
 
 import android.app.Activity
 import android.os.Bundle
-import android.support.v4.app.FragmentActivity
-import android.support.v7.app.AppCompatActivity
+import androidx.fragment.app.FragmentActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 
 /**
