@@ -1,6 +1,6 @@
 package com.example.yukai.mydefinedview.BasicView.RecyclerView;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
