@@ -1,6 +1,6 @@
 package com.example.yukai.mydefinedview.BasicView.ViewPager;
 
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.view.View;
 
 /**
