@@ -3,7 +3,7 @@ package com.example.yukai.mydefinedview.ExampleTest.AsyncTaskTest;
 import android.app.Activity;
 import android.graphics.Bitmap;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
