@@ -3,7 +3,7 @@ package com.example.yukai.mydefinedview.ExternalView;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 
 import com.example.yukai.mydefinedview.ExternalView.BiliBiliDanmuView.BiliBiliDanmuViewActivity;
