@@ -2,10 +2,9 @@ package com.example.yukai.mydefinedview.MyView.MyAutoLayout;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.example.yukai.mydefinedview.R;
