@@ -2,7 +2,7 @@ package com.example.yukai.mydefinedview.WorkTest.FullRoomUITest;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.example.yukai.mydefinedview.R;
 
