@@ -2,7 +2,7 @@ package com.example.yukai.mydefinedview.WorkTest.TouchEventTest;
 
 import android.content.Context;
 import android.graphics.Color;
-import android.support.v4.view.PagerAdapter;
+import androidx.viewpager.widget.PagerAdapter;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
