@@ -2,8 +2,8 @@ package com.example.yukai.mydefinedview.WorkTest.TouchEventTest;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.view.ViewPager;
+import androidx.annotation.Nullable;
+import androidx.viewpager.widget.ViewPager;
 
 import com.example.yukai.mydefinedview.R;
 
