@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.fragment.app.FragmentManager
 import com.example.yukai.mydefinedview.R
 
 /**
@@ -15,28 +16,28 @@ class Fragment3 : BaseFragment(){
     var name:String = ""
 
     override fun getLogName(): String {
-        return "Fragment1"
+        return "Fragment3"
     }
 
-    override fun onCreateView(inflater: LayoutInflater?, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        return inflater?.inflate(R.layout.fragment1_layout, container, false)
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        return inflater.inflate(R.layout.fragment1_layout, container, false)
     }
 
-    override fun onViewCreated(view: View?, savedInstanceState: Bundle?) {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         val textView = view?.findViewById(R.id.fragment_text) as? TextView
         textView?.text = "this is fragment3"
-        textView?.setOnClickListener(object : View.OnClickListener{
-            override fun onClick(v: View?) {
-                activity.supportFragmentManager.popBackStack("f2", android.support.v4.app.FragmentManager.POP_BACK_STACK_INCLUSIVE)
-            }
-        })
+        textView?.setOnClickListener {
+            activity?.supportFragmentManager?.popBackStack(
+                "f2",
+                FragmentManager.POP_BACK_STACK_INCLUSIVE
+            )
+        }
     }
 
     companion object{
-        fun newInstance(name: String): Fragment1{
-            val fragment = Fragment1()
-            return fragment
+        fun newInstance(name: String): Fragment1 {
+            return Fragment1()
         }
     }
 
