@@ -1,13 +1,13 @@
 package com.example.yukai.mydefinedview.review;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.example.yukai.mydefinedview.R;
 import com.example.yukai.mydefinedview.review.IntentService.ReviewSimpleActivity;
-import com.example.yukai.mydefinedview.review.MVVM.MvvmActivity;
 
 public class ReviewActivity extends AppCompatActivity implements View.OnClickListener {
 
@@ -29,8 +29,8 @@ public class ReviewActivity extends AppCompatActivity implements View.OnClickLis
                 startActivity(intent);
                 break;
             case R.id.btn_mvvm:
-                intent = new Intent(ReviewActivity.this, MvvmActivity.class);
-                startActivity(intent);
+//                intent = new Intent(ReviewActivity.this, MvvmActivity.class);
+//                startActivity(intent);
                 break;
         }
     }
